separate UpdateType definition into its own file
