+++ backend/src/document-utils/realtimeDocumentUpdates.ts
@@ -1,5 +1,6 @@
 import { Document } from "@lib/src/Document";
-import { OnlineEntity, UpdateType } from "@lib/src/realtimeUserTypes";
+import { OnlineEntity } from "@lib/src/realtimeUserTypes";
+import { UpdateType } from "@lib/src/UpdateType";
 import { UserEntity } from "@lib/src/UserEntity";
 
 import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
