+++ backend/src/document-utils/realtimeOnlineUsers.ts
@@ -1,4 +1,6 @@
-import { OnlineEntity, UpdateType } from "@lib/src/realtimeUserTypes";
+import { OnlineEntity } from "@lib/src/realtimeUserTypes";
+import { UpdateType } from "@lib/src/UpdateType";
+
 import { UserEntity } from "@lib/src/UserEntity";
 
 import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
